Network Response
Addition of throwing null argument in the connecting with a GET.  At the moment it is assumed that the response is fully contained within the response of a GET request.

API modification for better naming and comments have been added.

Addition of entry level test case to check for NULL.