Network Utils
Removal of the URL API specific information from the NetworkUtils class.  This class will now only perform the GET operation that is necessary to receive the response from the provided URL.

The original URL builder has been moved to a new class known as the MoiveDataModel.

MovieDataModel is responsible for the follow
- constructing the URL from a set of pre-defined operations.
- parsing the response to a proper usable construct.

These will be added in later features.