Movie DB URL Construction
Popular and Top Rated URL building methods now support indicating a page if desired as well.

There was a minor bug with additional paths to the parsed base path.  Additional paths must be encoded otherwise the special characters might be evaluated in-correctly.

Also added test file for the model, however requires mocking infrastructure that will be ignored for the time being.