Image Displaying
The recycler view can now display images.
- The bind method within the ViewHolder of the Movie adapter class uses picasso to fetch the images.
- The path url has been adjusted within the JsonUtils class for the Movie Poster path.

At the moment only the first page is being viewed.

At the moment no progress bar.

At the moment during the activity the entire recycler view will disappear.

At the moment getting to the end of the RecyclerView position does not query for more data, will not want to query for all the pages if already have.