MovieAsyncTaskLoader
The usage of `IntentService` does not fit in well with the intended functionality of the application.
- `IntentService` should be used when background operations are needed but are not tied to the UI (i.e. does not deliver data back to UI).
- `AsyncTask` should be used when background operations are needed and need to return data back for UI functionality.

`AsyncTasks` by themselves are not desirable since upon a life-cycle transition to another non-active, state the task would deliver the data to a non-available instance, could crash the application.  Additionally this could cause several un-necessary zombie threads that keep getting spawned.

`AsyncTaskLoader` is being used.  This commit contains the logic for actually performing the work.  However it still needs to be associated with a task manager in order to deliever results to a possible new instance of the application activity.