Smoother Scrolling & Fetch of Data
Removed the available interface from the `MovieAdapter`.  It no longer supports a fetch data interface.  The adapter has no understanding of when to actually request data which is okay now.

Fetching of data is completed by the main activity, and will set the data which will notify the adapter that the data set has changed by calling the `setMovies` method in the adapter class.

Fetching of data only happens one at a time even if it fails failure would expect null data to be sent back of some form.

At the moment this is unbounded growth for number of movies associated with the application.