Clean Up
Main Activity documentation and handling of multiple data sets now.
- new POJO has been added `MovieManagedData` to account for the data that will be populated from the async tasks loader at the end of the day.
- the actual async task does not actually manage this information but rather is handled by the LoaderCallBack LoaderManager instead within MainActivity.

The active data set that is being operated upon is available to the adapter, but primarily to be referenced during the scrolling event.  Granted this could be handled without this as well since `MainActivity` knows what the current active data set is.  Will likely remove later.