Detail Activity
Support for the movie details are passed to the intent.

Backdrop image is used for the background with some transparency.  It is purposefully skewed for the other elements on the activity will break the view apart.  Just being used to provide some unique coloring that is respective to the palatte associated with the movie.