Parcelable Movie Support
In order to support instance information, all the movie attributes that have already been attained and the settings will need to be saved in a bundle within the `onSaveInstance` method.

This commit supports the Movie `parcelable implementation that allows a `Movie` object to be saved (by way of breaking up its class attributes into the proper types).